refactor(frontend): extract TokenInputContent from TokenInput
Split the inner content (title, amount input strip, token selector,
amount/balance footer) of TokenInput into a standalone TokenInputContent
component. TokenInput now delegates to it inside its bordered card, so all
existing consumers render identically.

Use TokenInputContent directly in the limit-order trade pair box, which
already provides its own container, dropping the redundant inner card.
